Extract `Experience`
