+++ src/components/sections/Experience/Experience.module.scss
@@ -1,0 +1,58 @@
+.container {
+    width: 100%;
+  
+    li {
+      list-style-type: none;
+      position: relative;
+      display: flex;
+      gap: 0.8rem;
+  
+      &::before {
+        content: "•";
+        height: 21px;
+        position: absolute;
+        left: -8px;
+        display: flex;
+        align-items: center;
+  
+        @media screen and (min-width: $media-sm) {
+          top: 48%;
+          translate: 0 -50%;
+          left: -12px;
+        }
+      }
+  
+      > *:last-child {
+        margin-left: auto;
+      }
+  
+      @media screen and (min-width: $media-sm) {
+        flex-direction: row;
+      }
+    }
+  }
+  
+  .desktop {
+    display: none;
+  
+    @media screen and (min-width: $media-sm) {
+      display: block;
+    }
+  }
+  
+  .mobile {
+    @media screen and (min-width: $media-sm) {
+      display: none;
+    }
+  }
+  
+  .details {
+    display: flex;
+    justify-content: flex-start;
+    gap: 0.4rem;
+  
+    @media screen and (min-width: $media-sm) {
+      gap: 0.8rem;
+    }
+  }
+
