+++ src/components/sections/Experience/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Experience'
